moved script into single file, for now.
fixes bug that didn't display correct line.
